Prevent `computeOptimalChunks` from hanging onto the main thread
